# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a bookstore server required for Hypermedia Project by Politecncico\
    \ di Milano."
  version: "1.0.0"
  title: "Bookstore"
  contact:
    email: "gabriele.gallotti@mail.polimi.it"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/v2"
...
schemes:
- "http"
- "https"
paths:
...
  /user/login:
    post:
      tags:
      - "user"
      summary: "Login"
      description: "Login with a form"
      operationId: "POSTuserLogin"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "formData"
        required: true
        type: "string"
      - name: "password"
        in: "formData"
        required: true
        type: "string"
      responses:
        200:
          description: "succesfull login"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /user/register:
    post:
      tags:
      - "user"
      summary: "Register"
      description: "Register into the store"
      operationId: "POSTuserRegister"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "succesfull registration"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /user/{id}:
    get:
      tags:
      - "user"
      summary: "Find user by ID"
      description: "Returns a single user"
      operationId: "getUserById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of user to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "User"
  /user/{id}/orders:
    get:
      tags:
      - "user"
      summary: "retrieves the orders of an user"
      operationId: "getOrdersByUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the user to retrieve the orders from"
        required: true
        type: "integer"
        format: "int64"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Order"
      x-swagger-router-controller: "User"
  /user/{id}/reviews:
    get:
      tags:
      - "user"
      summary: "retrieves the reviews of an user"
      operationId: "getReviewsByUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the user to retrieve the reviews from"
        required: true
        type: "integer"
        format: "int64"
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Review"
      x-swagger-router-controller: "User"
  /user/{id}/cart:
    get:
      tags:
      - "user"
      summary: "retrieves the cart of an user"
      operationId: "getCartByUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the user to retrieve the cart from"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Cart"
      x-swagger-router-controller: "User"
...
definitions:
  User:
    type: "object"
    required:
    - "id"
    - "name"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      address:
        type: "string"
      creditcard:
        type: "string"
    title: "User"
    description: "A user"
    example:
      id: 1
      name: "Davide"
      address: "Via Roma, Roma"
      creditcard: "xdfehc"
  Cart:
    type: "object"
    properties:
      total:
        $ref: "#/definitions/Amount"
      books:
        type: "array"
        items:
          $ref: "#/definitions/Book"
    title: "Cart"
    description: "Order for books"
    example:
      total:
        currency: "eur"
        value: 80082819046101.171875
      books:
      - id: 0
        title: "1984"
        author: "Orwell George"
        price:
          value: 12
          currency: "eur"
      - id: 1
        title: "Guida Galattica per autostoppisti"
        author: "Douglas Adams"
        price:
          value: 10
          currency: "eur"
  Book:
    type: "object"
    required:
    - "author"
    - "id"
    - "price"
    - "title"
    properties:
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      review:
        type: "array"
        items:
          $ref: "#/definitions/Review"
      author:
        type: "array"
        items:
          $ref: "#/definitions/Author"
      price:
        $ref: "#/definitions/Amount"
      eBook:
        type: "boolean"
        description: "True if eBook is available"
      intro:
        type: "string"
      genre:
        type: "string"
      factsheet:
        type: "string"
      abstract:
        type: "string"
      firstTheme:
        type: "string"
      secondTheme:
        type: "string"
      thirdTheme:
        type: "string"
      events:
        type: "array"
        items:
          $ref: "#/definitions/Event"
      status:
        type: "string"
        description: "book availability in the inventory"
        enum:
        - "Available"
        - "out of stock"
    title: "Book"
    description: "A book for sale in the store"
    example:
      id: 0
      title: "Brave new world"
      author: "Aldous Huxley"
      price:
        value: 10
        currency: "eur"
      status: "available"
  Review:
    type: "object"
    required:
    - "book"
    - "stars"
    - "user"
    properties:
      user:
        $ref: "#/definitions/User"
      book:
        $ref: "#/definitions/Book"
      stars:
        type: "number"
        format: "integer"
        description: "vote to the book from 1 to 5"
        minimum: 1
        maximum: 5
      review:
        type: "string"
        description: "text of the review"
    example:
      user:
        name: "Davide"
      book:
        id: 0
        title: "1984"
        author: "Orwell George"
        price:
          value: 12
          currency: "eur"
      stars: 2
  Event:
    type: "object"
    required:
    - "date"
    - "id"
    - "overview"
    - "place"
    properties:
      authors:
        type: "array"
        items:
          $ref: "#/definitions/Author"
      books:
        type: "array"
        items:
          $ref: "#/definitions/Book"
      id:
        type: "integer"
        format: "int64"
      overview:
        format: "string"
      place:
        format: "string"
      date:
        format: "date"
      advertisingPoster:
        format: "jpg"
      award:
        format: "string"
    example:
      date: ""
      overview: ""
      books:
      - id: 0
        title: "Brave new world"
        author: "Aldous Huxley"
        price:
          value: 10
          currency: "eur"
        status: "available"
      - id: 0
        title: "Brave new world"
        author: "Aldous Huxley"
        price:
          value: 10
          currency: "eur"
        status: "available"
      award: ""
      advertisingPoster: ""
      id: 0
      place: ""
      authors:
      - books:
        - id: 0
          title: "Brave new world"
          author: "Aldous Huxley"
          price:
            value: 10
            currency: "eur"
          status: "available"
        - id: 0
          title: "Brave new world"
          author: "Aldous Huxley"
          price:
            value: 10
            currency: "eur"
          status: "available"
        award: ""
        name: ""
        photo: ""
        id: 0
        life: ""
      - books:
        - id: 0
          title: "Brave new world"
          author: "Aldous Huxley"
          price:
            value: 10
            currency: "eur"
          status: "available"
        - id: 0
          title: "Brave new world"
          author: "Aldous Huxley"
          price:
            value: 10
            currency: "eur"
          status: "available"
        award: ""
        name: ""
        photo: ""
        id: 0
        life: ""
  Order:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      userId:
        type: "integer"
        format: "int64"
      books:
        type: "array"
        items:
          $ref: "#/definitions/Book"
      creationDate:
        type: "string"
        format: "date-time"
    example:
      books:
      - id: 0
        title: "Brave new world"
        author: "Aldous Huxley"
        price:
          value: 10
          currency: "eur"
        status: "available"
      - id: 0
        title: "Brave new world"
        author: "Aldous Huxley"
        price:
          value: 10
          currency: "eur"
        status: "available"
      id: 0
      creationDate: "2000-01-23T04:56:07.000+00:00"
      userId: 6
  Author:
    type: "object"
    required:
    - "books"
    - "id"
    - "name"
    properties:
      books:
        type: "array"
        items:
          $ref: "#/definitions/Book"
      id:
        type: "integer"
        format: "int64"
      name:
        format: "string"
      life:
        format: "string"
      photo:
        format: "jpg"
      award:
        format: "string"
    example:
      books:
      - id: 0
        title: "Brave new world"
        author: "Aldous Huxley"
        price:
          value: 10
          currency: "eur"
        status: "available"
      - id: 0
        title: "Brave new world"
        author: "Aldous Huxley"
        price:
          value: 10
          currency: "eur"
        status: "available"
      award: ""
      name: ""
      photo: ""
      id: 0
      life: ""
...
  Amount:
    type: "object"
    required:
    - "currency"
    - "value"
    properties:
      value:
        type: "number"
        format: "double"
        minimum: 0.01
        maximum: 10000
      currency:
        $ref: "#/definitions/Currency"
    description: "Price\n"
    example:
      currency: "eur"
      value: 11
  Currency:
    type: "string"
    pattern: "^[A-Z]{3,3}$"
    description: "some description\n"
    example: "eur"
...